Add updated font
