+++ app/src/main/java/com/cornellappdev/scoop/ui/theme/Type.kt
@@ -1,0 +1,35 @@
+package com.cornellappdev.scoop.ui.theme
+
+import androidx.compose.material.Typography
+import androidx.compose.ui.text.TextStyle
+import androidx.compose.ui.text.font.Font
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.sp
+import com.cornellappdev.scoop.R
+
+private val Roboto = FontFamily(
+    Font(R.font.roboto_regular),
+    Font(R.font.roboto_light, FontWeight.Light),
+    Font(R.font.roboto_bold, FontWeight.Bold),
+    Font(R.font.roboto_medium, FontWeight.Medium),
+    Font(R.font.roboto_bold, FontWeight.Bold),
+    Font(R.font.roboto_black, FontWeight.Black)
+)
+
+// Set of Material typography styles to start with
+val Typography = Typography(
+    defaultFontFamily = Roboto,
+    body1 = TextStyle(
+        fontFamily = FontFamily.Default,
+        fontWeight = FontWeight.Normal,
+        fontSize = 16.sp
+    ),
+    h4 = TextStyle(
+        fontWeight = FontWeight.W700,
+        fontSize = 34.sp
+    ),
+    h5 = TextStyle(
+        fontSize = 22.sp
+    )
+)
